feat: deprecate imageHelper mixin
