+++ frontend/components/Item/Card/Card.vue
@@ -85,13 +85,12 @@
 import Vue from 'vue';
 import { mapStores } from 'pinia';
 import { BaseItemDto, ImageType } from '@jellyfin/client-axios';
-import imageHelper from '~/mixins/imageHelper';
 import itemHelper from '~/mixins/itemHelper';
 import { CardShapes, getShapeFromItemType } from '~/utils/items';
 import { socketStore } from '~/store';
 
 export default Vue.extend({
-  mixins: [imageHelper, itemHelper],
+  mixins: [itemHelper],
   props: {
     item: {
       type: Object as () => BaseItemDto,
@@ -256,11 +255,7 @@ export default Vue.extend({
   },
   methods: {
     isFinePointer(): boolean {
-      if (process.client) {
-        return window.matchMedia('(pointer:fine)').matches;
-      } else {
-        return false;
-      }
+      return window.matchMedia('(pointer:fine)').matches;
     }
   }
 });
